+++ src/main/java/io/github/thebusybiscuit/slimefun4/core/networks/energy/EnergyNet.java
@@ -1,6 +1,8 @@
 package io.github.thebusybiscuit.slimefun4.core.networks.energy;
 
+import java.util.HashMap;
 import java.util.HashSet;
+import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
 import java.util.concurrent.atomic.AtomicLong;
@@ -15,6 +17,7 @@ import io.github.thebusybiscuit.slimefun4.api.ErrorReport;
 import io.github.thebusybiscuit.slimefun4.api.network.Network;
 import io.github.thebusybiscuit.slimefun4.api.network.NetworkComponent;
 import io.github.thebusybiscuit.slimefun4.core.attributes.EnergyNetComponent;
+import io.github.thebusybiscuit.slimefun4.core.attributes.EnergyNetProvider;
 import io.github.thebusybiscuit.slimefun4.implementation.SlimefunItems;
 import io.github.thebusybiscuit.slimefun4.implementation.SlimefunPlugin;
 import io.github.thebusybiscuit.slimefun4.implementation.items.electric.reactors.Reactor;
@@ -35,6 +38,7 @@ import me.mrCookieSlime.Slimefun.api.energy.ChargableBlock;
  * 
  * @see Network
  * @see EnergyNetComponent
+ * @see EnergyNetProvider
  * @see EnergyNetComponentType
  *
  */
@@ -49,16 +53,10 @@ public class EnergyNet extends Network {
             return EnergyNetComponentType.NONE;
         }
 
-        if (SlimefunPlugin.getRegistry().getEnergyGenerators().contains(id)) {
-            return EnergyNetComponentType.GENERATOR;
-        }
+        SlimefunItem item = SlimefunItem.getByID(id);
 
-        if (SlimefunPlugin.getRegistry().getEnergyCapacitors().contains(id)) {
-            return EnergyNetComponentType.CAPACITOR;
-        }
-
-        if (SlimefunPlugin.getRegistry().getEnergyConsumers().contains(id)) {
-            return EnergyNetComponentType.CONSUMER;
+        if (item instanceof EnergyNetComponent) {
+            return ((EnergyNetComponent) item).getEnergyComponentType();
         }
 
         return EnergyNetComponentType.NONE;
@@ -143,10 +141,10 @@ public class EnergyNet extends Network {
             SimpleHologram.update(b, "&4No Energy Network found");
         }
         else {
-            double supply = DoubleHandler.fixDouble(tickAllGenerators(timestamp::getAndAdd) + tickAllCapacitors());
-            double demand = 0;
-
-            int availableEnergy = (int) supply;
+            Map<Location, Integer> generatorsWithCapacity = new HashMap<>();
+            int supply = tickAllGenerators(generatorsWithCapacity, timestamp::getAndAdd) + tickAllCapacitors();
+            int remainingEnergy = supply;
+            int demand = 0;
 
             for (Location machine : consumers) {
                 int capacity = ChargableBlock.getMaxCharge(machine);
@@ -156,20 +154,20 @@ public class EnergyNet extends Network {
                     int availableSpace = capacity - charge;
                     demand += availableSpace;
 
-                    if (availableEnergy > 0) {
-                        if (availableEnergy > availableSpace) {
+                    if (remainingEnergy > 0) {
+                        if (remainingEnergy > availableSpace) {
                             ChargableBlock.setUnsafeCharge(machine, capacity, false);
-                            availableEnergy -= availableSpace;
+                            remainingEnergy -= availableSpace;
                         }
                         else {
-                            ChargableBlock.setUnsafeCharge(machine, charge + availableEnergy, false);
-                            availableEnergy = 0;
+                            ChargableBlock.setUnsafeCharge(machine, charge + remainingEnergy, false);
+                            remainingEnergy = 0;
                         }
                     }
                 }
             }
 
-            storeExcessEnergy(availableEnergy);
+            storeExcessEnergy(generatorsWithCapacity, remainingEnergy);
             updateHologram(b, supply, demand);
         }
 
@@ -177,7 +175,7 @@ public class EnergyNet extends Network {
         SlimefunPlugin.getProfiler().closeEntry(b.getLocation(), SlimefunItems.ENERGY_REGULATOR.getItem(), timestamp.get());
     }
 
-    private void storeExcessEnergy(int available) {
+    private void storeExcessEnergy(Map<Location, Integer> generators, int available) {
         for (Location capacitor : storage) {
             if (available > 0) {
                 int capacity = ChargableBlock.getMaxCharge(capacitor);
@@ -196,39 +194,81 @@ public class EnergyNet extends Network {
             }
         }
 
-        for (Location generator : generators) {
-            int capacity = ChargableBlock.getMaxCharge(generator);
+        for (Map.Entry<Location, Integer> entry : generators.entrySet()) {
+            Location generator = entry.getKey();
+            int capacity = entry.getValue();
 
-            if (capacity > 0) {
-                if (available > 0) {
-                    if (available > capacity) {
-                        ChargableBlock.setUnsafeCharge(generator, capacity, false);
-                        available = available - capacity;
-                    }
-                    else {
-                        ChargableBlock.setUnsafeCharge(generator, available, false);
-                        available = 0;
-                    }
+            if (available > 0) {
+                if (available > capacity) {
+                    ChargableBlock.setUnsafeCharge(generator, capacity, false);
+                    available -= capacity;
                 }
                 else {
-                    ChargableBlock.setUnsafeCharge(generator, 0, false);
+                    ChargableBlock.setUnsafeCharge(generator, available, false);
+                    available = 0;
                 }
+            }
+            else {
+                ChargableBlock.setUnsafeCharge(generator, 0, false);
             }
         }
     }
 
-    private double tickAllGenerators(LongConsumer timeCallback) {
-        double supply = 0;
+    private int tickAllGenerators(Map<Location, Integer> generatorsWithCapacity, LongConsumer timeCallback) {
         Set<Location> exploded = new HashSet<>();
+        int supply = 0;
 
         for (Location source : generators) {
             long timestamp = SlimefunPlugin.getProfiler().newEntry();
             Config config = BlockStorage.getLocationInfo(source);
             SlimefunItem item = SlimefunItem.getByID(config.getString("id"));
 
-            if (item != null) {
+            if (item instanceof EnergyNetProvider) {
                 try {
+                    EnergyNetProvider provider = (EnergyNetProvider) item;
+                    int energy = provider.getGeneratedOutput(source, config);
+
+                    if (provider.getCapacity() > 0) {
+                        generatorsWithCapacity.put(source, provider.getCapacity());
+                        String charge = config.getString("energy-charge");
+
+                        if (charge != null) {
+                            energy += Integer.parseInt(charge);
+                        }
+                    }
+
+                    if (provider.willExplode(source, config)) {
+                        exploded.add(source);
+                        BlockStorage.clearBlockInfo(source);
+
+                        Slimefun.runSync(() -> {
+                            source.getBlock().setType(Material.LAVA);
+                            source.getWorld().createExplosion(source, 0F, false);
+                        });
+                    }
+                    else {
+                        supply += energy;
+                    }
+                }
+                catch (Exception | LinkageError t) {
+                    exploded.add(source);
+                    new ErrorReport(t, source, item);
+                }
+
+                long time = SlimefunPlugin.getProfiler().closeEntry(source, item, timestamp);
+                timeCallback.accept(time);
+            }
+            else if (item != null) {
+                try {
+                    // This will be removed very very soon
+                    // It is only here for temporary backwards compatibility
                     GeneratorTicker generator = item.getEnergyTicker();
+
+                    Integer capacity = SlimefunPlugin.getRegistry().getEnergyCapacities().get(item.getID());
+
+                    if (capacity != null && capacity.intValue() > 0) {
+                        generatorsWithCapacity.put(source, capacity);
+                    }
 
                     if (generator != null) {
                         double energy = generator.generateEnergy(source, item, config);
@@ -267,12 +307,11 @@ public class EnergyNet extends Network {
         }
 
         generators.removeAll(exploded);
-
         return supply;
     }
 
-    private double tickAllCapacitors() {
-        double supply = 0;
+    private int tickAllCapacitors() {
+        int supply = 0;
 
         for (Location capacitor : storage) {
             supply += ChargableBlock.getCharge(capacitor);